✅ test: test app 생성 fixture로 변경
