+++ server/test/jest-integration.json
@@ -6,5 +6,6 @@
   "transform": {
     "^.+\\.(t|j)s$": "ts-jest"
   },
-  "coverageDirectory": "test/coverage"
+  "coverageDirectory": "test/coverage",
+  "setupFilesAfterEnv": ["./test/jest.setup.ts"]
 }
